requirements.txt
openpyxl
Pillow
pyperclip
tkinterdnd2
